Move package source to src, tests to tests
