+++ .gitignore
@@ -1,9 +1,8 @@
-/.gitattributes
 /.vscode/
 
 /build/
 /dist/
-/dicognito.egg-info/
+/src/dicognito.egg-info/
 /smoketest/*.dcm
 /smoketest/*.txt
