+++ src/dicognito/__init__.py
@@ -1,0 +1,1 @@
+'''dicognito - anonymize DICOM datasets'''
